project_db tests: move DDL tests into test_schema.py
